ai-ppdb/ppdb/test/black/requirements.txt
python-memcached>=1.59 
pymemcache>=4.0.0